money/robinhood_bot/requirements.txt
python-dotenv
openai
discord-webhook
robinhood-crypto
pandas
requests
ta-lib
matplotlib
typer
